+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# xenoblade_blender Changelog
+# sm4sh_blender Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).